Portal/backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
PyJWT==2.8.0
SQLAlchemy==2.0.23
PyMySQL==1.1.0
cryptography==41.0.7
PyMuPDF==1.24.0
reportlab==4.1.0
Pillow==10.2.0
anthropic==0.40.0
python-dotenv==1.0.0
pyotp==2.9.0
bcrypt==4.1.2
